Mobile debugging
Updated skills section as SwiperJS component/slides did not have a set height.

The Timeline + items needed to be altered for browser thrashing and slowing down animations on both desktop and mobile as well as expanding the timeline so the user can scroll and experience the animation.

The Navbar for mobile was modified included a display of flex and padding 0 0 0 0 to ensure the navbar buttons sit properly and take up mobile device width.